# builtin.string: optimize split_into_lines